Decode first 5 bits of thumb instruction
For most thumb instructions, the first 5 bits or less are opcodes that identify the instruction, with a few exceptions such as the condition branch instructions. While the solution may not be so elegant, it compiles to a jumptable of 31 labels, striking a balance between efficiency and convenience.

While implementing a hash function will be more efficient, suspections are that it will consums too much memory and hurt cache locality. This remains an alternative to be tested in the future.

Now the program can disassemble most of the test rom file in thumb mode roughly.