Model memory and finish disassembling of data processing with 8-bit immediate operand.
The memory is modeled as a struct consisting of several Vec<u8>s. Different Vec<u8> represent different fields of memory, for instance, BIOS, internal/external work ram, etc. The reason to ruse Vec<u8> instead of [u8] is that arrays are allocated on the stack, and any data larger than 8 MiB will cause the stack to overflow. Another alternative remains to be tested is to use Box<[u8]>. Data processing part of the disassemble has some redundancy.