Implement getters and setters for `cpsr`, and prototype debugger.
`cpsr` consists of condition code flags and control bits. Directly modifying it is dangerous and will probably worsen readability. Thus field `cpsr` and `spsr` of `Register` are made private and `get/set` functions are added. Currently `debug.rs` only have one `debug()` function, a struct and methods are to be added in the future.