Implement conditional execution, improve debugging, and add new `Console` struct.
Conditional execution is implemented by `check_condition()`, which checks CPSR flags accordingly, returning a boolean. Note that `check_condition()` is places inside `cpu.rs` because thumb mode still need it for branching operations. `step`, `print`, and `disassemble` primitives are added to `debug.rs`. As the debugger needs to access the CPU and memory at the same time, a new `Console` composing the two are added. `timer`, `lcd`, `speaker` and other components are yet to be added and implement.