Implement mode switch, fix pc related bugs
Banked registers are all stored in the `bank` field of struct `CPU`. Because FIQ mode has more banked registers than other modes, an ad hoc approach is taken to cope with it.
Manipulation of R15 must be done with caution. Mysterious bugs are spotted when R15 is used as operands or destination in data processing.